Initial commit
This is the first commit (obviously). I took care of the project setup
as well as crate location parsing. If you look in the `/meta/` directory
in `/cratesystem/` you'll notice a class called SafeLocation.

This class was written as a result of an issue parsing the Bukkit
Location object to JSON with the Gson library. The server console throws
obscure errors such ass "field already exists with name d" which is
likely a result of the level of abstraction and pure chaos of the Bukkit
library (I had this problem parsing the ItemStack object in a past
project).

I figured that, instead of creating a whole system to help me serialize
the Location object to JSON, which would be a waste of time and space, I
can just store the important data from the Location object in
SafeLocation, and just parse that data to Bukkit Location when I need
to.

Thanks for the read!