Completed compiler
The compiler that takes care of parsing tables in
plugins/CrateSystem/tables/ is done. I have staretd writing
documentation on how to use the crate system. Expect future updates of
the docs. I want the system to be accessible to everyone so not
everybody who wants to make random minigame loot can just work off of
mine instead of having to do the entire thing themselves. The whole
point of development is to advance science for everybody, even if it
abstracting from the bare bones of net.minecraft.server a little bit.
Cheers!